Set step result on both success and failure
Right now there is no way to tell which step fails. This is particularly
helpful in the case of:
* Parallel steps where any one failure could fail the workflow
* Debugging flakey or low success rate jobs
* Jobs which have retries (automated or manual)

The `with_span` command works by running the call to Honeycomb in a
`when:always` wrapped step. To the best of my knowledge we do not know
the status of previous steps at this point, yet that is precisely what
we need to know to add that context.

In order to duplicate as little code as possible, I have introduced
independent `on_success` and `on_fail` steps which leverages the already
defined process of setting context via the `extra_fields.lgfmt` file.