Use single quoted bash strings
I believe there's an issue with the way this orb works in that it will overwrite your config. Using single quotes means the line is evaluated when the echo command is executed, not when the file is sourced. This broke our $PATH in our project, and we were able to fix it by manually resetting the path later in the flow.