Fix: Double quote vars that require interpolation
These three cases require interpolation of other environment variables. Use
of single quotes was not allowing this interpolation, and in the case of 
"BUILDEVENTS_SPAN_ID" the lack of this export causes issues with sending
spans.

Workaround is possible if you manually set the span ID like:
```
- run:
     name: Sleep
     command: |
     export BUILDEVENTS_SPAN_ID=$(echo ${CIRCLE_JOB}-${CIRCLE_NODE_INDEX} | sha256sum | awk '{print $1}')
     buildevents cmd ${CIRCLE_WORKFLOW_ID} ${BUILDEVENTS_SPAN_ID} 'run dependencies' -- sleep 5
```

NOTE: My blocker is on the span ID, however I have updated the quotes for the path exports as well. Also,
I am not sure how to test this as I am not sure how to get a test version of the orb without a commit to
the repo. Happy to do the leg work if there is a path for this, else please do check before merging! Thanks 🙇