ci: add CI integration test (#101)
In an attempt to implement a CI job that uses the orb itself to produce a real live Honeycomb trace, I seem to have refactored it into using orb-tools.

The only thing new in src/ is the example. Everything else is simply the old orb.yml split into pieces.

The actual integration test is in test-deploy.yml, which is the second step in our magical multi-part pipeline. It actually loads in the HEAD version of the orb and uses it to produce a trace.

Co-authored-by: Jason Harley <jharley@redmind.ca>